Fixed potential double conversion bug in buff ameter.
Corresponding to 912996f5f39e0704d79f72572c47de6e6d0ed483.

+++ src/haven/Buff.java
@@ -108,7 +108,7 @@ public class Buff extends Widget implements ItemInfo.ResOwner, Bufflist.Managed 
 
     public void draw(GOut g) {
 	g.chcolor(255, 255, 255, a);
-	Double ameter = (this.ameter >= 0) ? (this.ameter / 100.0) : ameteri.get();
+	Double ameter = (this.ameter >= 0) ? Double.valueOf(this.ameter / 100.0) : ameteri.get();
 	if(ameter != null) {
 	    g.image(cframe, Coord.z);
 	    g.chcolor(0, 0, 0, a);
